gas-oracle: L1 base fee max/min change
Add new config to prevent changes in the L1 base fee value
in the `OVM_GasPriceOracle` from being too volatile. New config
option `--max-percent-change-l1-base-fee` expects a float and
will cap upwards or downwards the amount that the base fee can
change between two consecutive updates. This is useful for preventing
too much volatility in the L1 portion of the fee that is paid on
L2.